Improved entity encoding
For ways and relations the ids are also stored in bulk now, then the
rest of the data, hence parsing just ids can be done more efficently.

We store the lengths of parts of the blocks in front of the actual data
(tag pools, ids, coordinates, members, tags, etc.) so that they can be
skipped by parsers that do not need some of the data.